Add resizing action
