+++ Document/Sources/Document/Drawing/Actions/MoveAction.swift
@@ -1,0 +1,36 @@
+import QuartzCore
+import AppKit
+
+public class MoveAction {
+    init(view: DrawingView, control: A11yControl, startLocation: CGPoint, offset: CGPoint, initialFrame: CGRect) {
+        self.view = view
+        self.control = control
+        self.startLocation = startLocation
+        self.offset = offset
+        self.initialFrame = initialFrame
+        
+        NSCursor.closedHand.push()
+    }
+    
+    private let view: DrawingView
+    public let control: A11yControl
+    private let startLocation: CGPoint
+    private(set) var offset: CGPoint
+    let initialFrame: CGRect
+    
+    public func drag(to coordinate: CGPoint) {
+        let offset = coordinate - startLocation
+        
+        let frame = initialFrame
+            .offsetBy(dx: offset.x,
+                      dy: offset.y)
+        
+        let aligned = view.alignmentOverlay.alignToAny(control, frame: frame, drawnControls: view.drawnControls)
+        
+        control.updateWithoutAnimation {
+            control.frame = aligned
+        }
+        
+        self.offset = offset
+    }
+}
